LAW LIB
I want to start to get the AI building Law libraries based of country, state or province. I live in Canada so I will start there.
it will also need to be tested on the best information layout for chatgpt to pull from. this is where rules and their association matter so that chat can take in infomation to see what laws my be applicable while still being able to reference laws should they be missed. finding out if rules are broken should enable a flag. a flag will bring up the rule and how to apply it.
I am likely not going to have a chance doing this task myself. this means it even needs a database for the paths of application of rules.